requirements.txt
django==4.0.3
djangorestframework
pyyaml
requests
django-cors-headers